drawio/requirements.txt
# Core ML and Computer Vision (CPU + GPU support)
torch>=1.10.0
torchvision>=0.11.0
# detectron2 - install manually based on your CUDA version:
# CPU only: pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch1.10/index.html
# CUDA 11.3: pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html
# For other versions, check: https://detectron2.readthedocs.io/en/latest/tutorials/install.html

opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Image Processing and OCR engines (multiple options for flexibility)
pillow>=8.3.0

# OCR Option 1: Tesseract (lightweight, CPU-friendly)
pytesseract>=0.3.8
# NOTE: Also install Tesseract binary:
# - Windows: https://github.com/UB-Mannheim/tesseract/wiki
# - Ubuntu: sudo apt-get install tesseract-ocr
# - Mac: brew install tesseract

# OCR Option 2: EasyOCR (good accuracy, supports many languages)
easyocr>=1.6.0

# OCR Option 3: PaddleOCR (best accuracy but may conflict)
# Uncomment if you want to use PaddleOCR (may cause conflicts):
# paddlepaddle>=2.4.0
# paddleocr>=2.6.0

# Data Processing
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Document Generation
python-pptx>=0.6.21
openpyxl>=3.0.9

# Progress and Utilities
tqdm>=4.62.0
requests>=2.25.0

# Development and Testing (optional)
pytest>=6.2.0
black>=21.6.0
flake8>=3.9.0

# XML Processing (for .drawio files)
lxml>=4.6.0

# Optional: Cloud OCR services (uncomment if needed)
# azure-cognitiveservices-vision-computervision>=0.9.0
# google-cloud-vision>=2.4.0
# boto3>=1.18.0  # for AWS Textract

# Optional: Alternative clustering
# hdbscan>=0.8.27

# Installation Tips:
# 1. For CPU-only (no GPU): Install CPU version of detectron2
# 2. For GPU: Install CUDA-compatible detectron2 version
# 3. OCR engines are listed in priority order (best → fallback)
# 4. You can install all OCR engines or pick one based on your needs
